LayoutTests/platform/chromium/fast/dom/Window/window-custom-prototype-expected.txt

Test what happens when you set the window's prototype to various values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL __proto__ = window; __proto should throw Error: cyclic __proto__ value. Threw exception Error: Cyclic __proto__ value.
FAIL __proto__ = chainPointingBackToWindow; __proto__ should throw Error: cyclic __proto__ value. Threw exception Error: Cyclic __proto__ value.
PASS __proto__ = 1; __proto__ is originalWindowPrototype
PASS __proto__ = 'a string'; __proto__ is originalWindowPrototype
PASS __proto__ = anotherObject; __proto__ is anotherObject
FAIL anotherObject.__proto__ = window; __proto__ should throw Error: cyclic __proto__ value. Threw exception Error: Cyclic __proto__ value.
PASS __proto__ = 1; __proto__ is anotherObject
PASS __proto__ = 'a string'; __proto__ is anotherObject
PASS __proto__ = anotherObject; __proto__ is anotherObject
PASS __proto__ = originalWindowPrototype; __proto__ is originalWindowPrototype
PASS anotherObject.__proto__ = window; anotherObject.__proto__ is window
FAIL __proto__ = null; __proto__ should be null. Threw exception ReferenceError: __proto__ is not defined
PASS successfullyParsed is true

TEST COMPLETE
